32-bit compilation fixes and cleanups

+++ storage/innobase/handler/ha_innodb.cc
@@ -16441,21 +16441,22 @@ wsrep_innobase_kill_one_trx(const trx_t *bf_trx, trx_t *victim_trx, ibool signal
 
 	if (!thd) {
 		DBUG_PRINT("wsrep", ("no thd for conflicting lock"));
-		WSREP_WARN("no THD for trx: %lu", victim_trx->id);
+		WSREP_WARN("no THD for trx: %llu", (long long)victim_trx->id);
 		DBUG_RETURN(1);
 	}
 	if (!bf_thd) {
 		DBUG_PRINT("wsrep", ("no BF thd for conflicting lock"));
-		WSREP_WARN("no BF THD for trx: %lu", (bf_trx) ? bf_trx->id : 0);
+		WSREP_WARN("no BF THD for trx: %llu",
+			   (bf_trx) ? (long long)bf_trx->id : 0);
 		DBUG_RETURN(1);
 	}
 
 	WSREP_LOG_CONFLICT(bf_thd, thd, TRUE);
 
-	WSREP_DEBUG("BF kill (%lu, seqno: %lld), victim: (%lu) trx: %lu",
+	WSREP_DEBUG("BF kill (%lu, seqno: %lld), victim: (%llu) trx: %llu",
  		    signal, (long long)bf_seqno,
- 		    wsrep_thd_thread_id(thd),
-		    victim_trx->id);
+ 		    (long long)wsrep_thd_thread_id(thd),
+		    (long long)victim_trx->id);
 
 	WSREP_DEBUG("Aborting query: %s", 
 		  (thd && wsrep_thd_query(thd)) ? wsrep_thd_query(thd) : "void");
@@ -16463,13 +16464,14 @@ wsrep_innobase_kill_one_trx(const trx_t *bf_trx, trx_t *victim_trx, ibool signal
 	wsrep_thd_LOCK(thd);
 
 	if (wsrep_thd_query_state(thd) == QUERY_EXITING) {
-		WSREP_DEBUG("kill trx EXITING for %lu", victim_trx->id);
+		WSREP_DEBUG("kill trx EXITING for %llu",
+			    (long long)victim_trx->id);
 		wsrep_thd_UNLOCK(thd);
 		DBUG_RETURN(0);
 	}
 	if(wsrep_thd_exec_mode(thd) != LOCAL_STATE) {
-		WSREP_DEBUG("withdraw for BF trx: %lu, state: %d",
-			    victim_trx->id,
+		WSREP_DEBUG("withdraw for BF trx: %llu, state: %d",
+			    (long long)victim_trx->id,
 		wsrep_thd_conflict_state(thd));
 	}
 
@@ -16478,8 +16480,8 @@ wsrep_innobase_kill_one_trx(const trx_t *bf_trx, trx_t *victim_trx, ibool signal
 		wsrep_thd_set_conflict_state(thd, MUST_ABORT);
 		break;
         case MUST_ABORT:
-		WSREP_DEBUG("victim %lu in MUST ABORT state", 
-			    victim_trx->id);
+		WSREP_DEBUG("victim %llu in MUST ABORT state",
+			    (long long)victim_trx->id);
 		wsrep_thd_UNLOCK(thd);
 		wsrep_thd_awake(thd, signal);
 		DBUG_RETURN(0);
@@ -16487,8 +16489,9 @@ wsrep_innobase_kill_one_trx(const trx_t *bf_trx, trx_t *victim_trx, ibool signal
 	case ABORTED:
 	case ABORTING: // fall through
 	default:
-		WSREP_DEBUG("victim %lu in state %d", 
-			    victim_trx->id, wsrep_thd_conflict_state(thd));
+		WSREP_DEBUG("victim %llu in state %d",
+			    (long long)victim_trx->id,
+			    wsrep_thd_conflict_state(thd));
 		wsrep_thd_UNLOCK(thd);
 		DBUG_RETURN(0);
 		break;
@@ -16498,8 +16501,8 @@ wsrep_innobase_kill_one_trx(const trx_t *bf_trx, trx_t *victim_trx, ibool signal
 	case QUERY_COMMITTING:
 		enum wsrep_status rcode;
 
-		WSREP_DEBUG("kill trx QUERY_COMMITTING for %lu", 
-			    victim_trx->id);
+		WSREP_DEBUG("kill trx QUERY_COMMITTING for %llu", 
+			    (long long)victim_trx->id);
 		wsrep_thd_awake(thd, signal); 
 
 		if (wsrep_thd_exec_mode(thd) == REPL_RECV) {
@@ -16513,8 +16516,8 @@ wsrep_innobase_kill_one_trx(const trx_t *bf_trx, trx_t *victim_trx, ibool signal
 			
 			switch (rcode) {
 			case WSREP_WARNING:
-				WSREP_DEBUG("cancel commit warning: %lu",
-					    victim_trx->id);
+				WSREP_DEBUG("cancel commit warning: %llu",
+					    (long long)victim_trx->id);
 				wsrep_thd_UNLOCK(thd);
 				DBUG_RETURN(1);
 				break;
@@ -16522,9 +16525,9 @@ wsrep_innobase_kill_one_trx(const trx_t *bf_trx, trx_t *victim_trx, ibool signal
 				break;
 			default:
 				WSREP_ERROR(
-					"cancel commit bad exit: %d %lu", 
+					"cancel commit bad exit: %d %llu", 
 					rcode, 
-					victim_trx->id);
+					(long long)victim_trx->id);
 				/* unable to interrupt, must abort */
 				/* note: kill_mysql() will block, if we cannot.
 				 * kill the lock holder first.
@@ -16538,7 +16541,7 @@ wsrep_innobase_kill_one_trx(const trx_t *bf_trx, trx_t *victim_trx, ibool signal
 		/* it is possible that victim trx is itself waiting for some 
 		 * other lock. We need to cancel this waiting
 		 */
-		WSREP_DEBUG("kill trx QUERY_EXEC for %lu", victim_trx->id);
+		WSREP_DEBUG("kill trx QUERY_EXEC for %llu", (long long)victim_trx->id);
 
 		victim_trx->lock.was_chosen_as_deadlock_victim= TRUE;
 		if (victim_trx->lock.wait_lock) {
@@ -16571,8 +16574,8 @@ wsrep_innobase_kill_one_trx(const trx_t *bf_trx, trx_t *victim_trx, ibool signal
 	{
 		bool skip_abort= false;
 		wsrep_aborting_thd_t abortees;
-                        
-		WSREP_DEBUG("kill IDLE for %lu", victim_trx->id);
+
+		WSREP_DEBUG("kill IDLE for %llu", (long long)victim_trx->id);
 
 		if (wsrep_thd_exec_mode(thd) == REPL_RECV) {
 			WSREP_DEBUG("kill BF IDLE, seqno: %lld",
